+++ src/main/java/com/rodrigo/br/gerenciador/repository/ResponsavelRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.rodrigo.br.gerenciador.modelo.Responsavel;
 
 @Repository
-public interface ResponsavelRepository extends JpaRepository<Responsavel, Integer> {
+public interface ResponsavelRepository extends JpaRepository<Responsavel, Long> {
 
     Responsavel findByNome(String nomeDoresponsavel);
