+++ src/main/java/com/rodrigo/br/gerenciador/repository/TarefaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.rodrigo.br.gerenciador.modelo.Tarefa;
 
-public interface TarefaRepository extends JpaRepository<Tarefa, Integer> {
+public interface TarefaRepository extends JpaRepository<Tarefa, Long> {
     
 }
